+++ MapboxNavigation/RouteMapViewController.swift
@@ -71,6 +71,8 @@ class RouteMapViewController: UIViewController {
         }
     }
     
+    var styleObservation: NSKeyValueObservation?
+    
     weak var delegate: RouteMapViewControllerDelegate?
     var routeController: Router! {
         didSet {
@@ -127,11 +129,16 @@ class RouteMapViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        let mapView = self.mapView
         mapView.contentInset = contentInsets
         view.layoutIfNeeded()
         
         mapView.tracksUserCourse = true
         
+        styleObservation = mapView.observe(\.style, options: .new) { [weak self] (mapView, change) in
+            self?.showRouteIfNeeded()
+            mapView.localizeLabels()
+        }
         
         distanceFormatter.numberFormatter.locale = .nationalizedCurrent
         
@@ -183,6 +190,7 @@ class RouteMapViewController: UIViewController {
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
         removeTimer()
+        styleObservation = nil
     }
 
     func resumeNotifications() {
@@ -395,14 +403,6 @@ class RouteMapViewController: UIViewController {
         mapView.altitude = altitude
     }
     
-    func mapView(_ mapView: MGLMapView, imageFor annotation: MGLAnnotation) -> MGLAnnotationImage? {
-        return navigationMapView(mapView, imageFor: annotation)
-    }
-    
-    func mapView(_ mapView: MGLMapView, viewFor annotation: MGLAnnotation) -> MGLAnnotationView? {
-        return navigationMapView(mapView, viewFor: annotation)
-    }
-    
     func notifyDidChange(routeProgress: RouteProgress, location: CLLocation, secondsRemaining: TimeInterval) {
         resetETATimer()
         updateETA()
@@ -587,28 +587,7 @@ extension RouteMapViewController: NavigationViewDelegate {
         delegate?.mapViewControllerDidDismiss(self, byCanceling: true)
     }
     
-    // MARK: MGLMapViewDelegate
-    func mapView(_ mapView: MGLMapView, regionDidChangeAnimated animated: Bool) {
-        var userTrackingMode = mapView.userTrackingMode
-        if let mapView = mapView as? NavigationMapView, mapView.tracksUserCourse {
-            userTrackingMode = .followWithCourse
-        }
-        if userTrackingMode == .none && !isInOverviewMode {
-            navigationView.wayNameView.isHidden = true
-        }
-    }
-    
-    func mapView(_ mapView: MGLMapView, didFinishLoading style: MGLStyle) {
-        // This method is called before the view is added to a window
-        // (if the style is cached) preventing UIAppearance to apply the style.
-        showRouteIfNeeded()
-        self.mapView.localizeLabels()
-        delegate?.mapView?(mapView, didFinishLoading: style)
-    }
-    
-    func mapViewDidFinishLoadingMap(_ mapView: MGLMapView) {
-        delegate?.mapViewDidFinishLoadingMap?(mapView)
-    }
+    // MARK: VisualInstructionDelegate
     
     func label(_ label: InstructionLabel, willPresent instruction: VisualInstruction, as presented: NSAttributedString) -> NSAttributedString? {
         return delegate?.label?(label, willPresent: instruction, as: presented)
@@ -1031,7 +1010,7 @@ fileprivate extension UIViewAnimationOptions {
         }
     }
 }
-@objc protocol RouteMapViewControllerDelegate: NavigationMapViewDelegate, MGLMapViewDelegate, VisualInstructionDelegate {
+@objc protocol RouteMapViewControllerDelegate: NavigationMapViewDelegate, VisualInstructionDelegate {
 
     func mapViewControllerDidOpenFeedback(_ mapViewController: RouteMapViewController)
     func mapViewControllerDidCancelFeedback(_ mapViewController: RouteMapViewController)